requirements.txt
numpy>=1.24.0
Pillow>=10.0.0  # For image processing 